pyproject.toml
[project]
name = "plov"
version = "0.1.0"
description = "We love plov"
authors = [
    { name = "Kir Tarasov", email = "tarasyonochek@yandex.ru" }
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
# PROD
dependencies = [
    "aiohttp>=3.11.16,<4.0.0",
    "celery>=5.5.1,<6.0.0",
    "django-elasticsearch-dsl>=8.0,<9.0",
    "django-htmx>=1.23.0,<2.0.0",
    "django>=5.2,<6.0",
    "djangorestframework>=3.16.0,<4.0.0",
    "faker>=37.1.0,<38.0.0",
    "pillow>=11.1.0,<12.0.0",
    "psycopg2>=2.9.10,<3.0.0",
    "python-decouple>=3.8,<4.0",
    "python-telegram-bot>=22.0,<23.0",
    "redis>=5.2.1,<6.0.0",
    "requests>=2.32.3,<3.0.0",
    "slugify>=0.0.1,<0.0.2",
    "transliterate>=1.10.2,<2.0.0",
]


[project.urls]
Homepage = "https://gitlab.crja72.ru/django/2025/spring/course/projects/team-4"
Documentation = "https://gitlab.crja72.ru/django/2025/spring/course/projects/team-4/-/blob/main/README.md"
Repository = "https://gitlab.crja72.ru/django/2025/spring/course/projects/team-4"


[tool.django]
settings_module = "plov.settings"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


# DEV
[tool.poetry.group.dev.dependencies]
black = "~25.1.0"
django-debug-toolbar = "~5.1.0"
django-extensions = "~3.2.3"
django-stubs = "~5.1.3"
django-types = "~0.20.0"
djangorestframework-types = "~0.9.0"
ipython = "~8.12.0"
isort = "~6.0.1"
mypy = "~1.15.0"
pre-commit = "~4.2.0"
psycopg2-binary = "~2.9.10"
types-requests = "~2.32.0.20250328"


# FLAKE8
[tool.poetry.group.flake8.dependencies]
flake8 = "^7.2.0"
flake8-pyproject = "^1.2.3"
# Code quality
flake8-clean-block = "^0.1.2"
flake8-eradicate = "^1.5.0"
flake8-implicit-str-concat = "^0.5.0"
flake8-print = "^5.0.0"
# Style
flake8-commas = "^4.0.0"
flake8-import-order = "^0.18.2"
flake8-quotes = "^3.4.0"
flake8-use-pathlib = "^0.3.0"
pep8-naming = "^0.14.1"


# TEST
[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-django = "^4.11.1"
pytest-xdist = "^3.5.0"


# SECURITY
[tool.poetry.group.security.dependencies]
bandit = "~1.8.3"


[tool.black]
target-version = ["py310", "py311", "py312"]
line-length = 120
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | venv
    | migrations
    | __pycache__
)/
'''


[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["core", "forum", "homepage", "interactions", "leaderboard", "mocklms", "reviews", "search", "stickers", "users", "users_status"]
known_third_party = ["django"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["migrations", "venv", ".venv"]
default_section = "THIRDPARTY"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true


[tool.flake8]
max-line-length = 120
ignore = ["E203", "E266", "E501", "W503", "I202"]
import-order-style = "google"
inline-quotes = "single"
application-import-names = ["core", "forum", "homepage", "interactions", "leaderboard", "mocklms", "reviews", "search", "stickers", "users", "users_status"]
exclude = [".git", ".venv", "venv", "*/migrations/*", "__pycache__", "*.pyc"]


[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
strict_optional = true
show_error_codes = true
mypy_path = 'plov'
plugins = ["mypy_django_plugin.main"]
exclude = '''
(?x)(
    .venv
    | /venv/
    | /migrations/
)
'''


[tool.django-stubs]
django_settings_module = "plov.settings"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "plov.settings"
python_files = "tests.py test_*.py *_tests.py"
addopts = "--cov=core --cov=forum --cov=homepage --cov=interactions --cov=leaderboard --cov=mocklms --cov=reviews --cov=search --cov=stickers --cov=users --cov=users_status --verbose"
